+++ CONTRIBUTING.md
@@ -2,7 +2,7 @@
 
 ## Required:
 
-- `node`
+- `node` (`>= 14.18.0`)
 - `yarn` (v1)
 
 ## Getting started
@@ -33,17 +33,8 @@ yarn dev
 
 ## Testing
 
-(Our testing solution is still a WIP. This section will change when we implement a better way to integrate `storybook` and `cypress`)
+Tests live in the `cypress/component` directory (a new test should be added for you when you scaffold a new component).
 
-We use `cypress` for both e2e and unit (component) tests -- make sure to place the test inside the appropriate sub-directory in the `cypress/` directory: e2e tests in `cypress/e2e` and unit tests in `cypress/component`.
+You can run the unit tests with `yarn test` without having to start the dev server. If you want to run tests in `open` mode, run `yarn test:open`.
 
-You can run the unit tests with `yarn test:component` without having a dev server already running.
-
-For e2e tests you have two options:
-
-- with the dev server running in one terminal, open another and run `yarn cy:run`
-- run `yarn test:e2e` -- this will run the dev server _and_ `cypress` for you, using `concurrently` to manage each process.
-
-When writing e2e tests, make sure you call `cy.visit()` with the appropriate path (the one you defined inside `src/App.tsx`).
-
-To run all tests you can run `yarn test:all`. With this command, code coverage is collected and reported by `istanbul` and `nyc` -- you can find this information inside the generated `coverage` folder after tests finish.
+To run all tests and collect code coverage (with `istanbul` and `nyc`), you can run `yarn test:coverage`. You can then find coverage information inside the generated `coverage` folder after tests finish.
