+++ package.json
@@ -14,9 +14,9 @@
     "build-storybook": "build-storybook",
     "cy:run": "cypress run",
     "cy:open": "cypress open",
-    "test:component": "cypress run --component",
-    "test:e2e": "concurrently -s first -k 'yarn storybook' 'yarn cy:run'",
-    "test:all": "yarn clear-coverage && yarn test:component && yarn test:e2e && yarn report:cy",
+    "test": "cypress run --component",
+    "test:open": "cypress open --component",
+    "test:coverage": "yarn clear-coverage && yarn test && yarn report:cy",
     "report:cy": "nyc report --reporter lcov --reporter text --report-dir coverage/coverage-cypress --check-coverage",
     "clear-coverage": "rm -rf .nyc_output && rm -rf coverage",
     "prepare": "husky install",
@@ -35,6 +35,7 @@
   "devDependencies": {
     "@cypress/code-coverage": "^3.10.0",
     "@cypress/react": "^5.12.5",
+    "@cypress/vite-dev-server": "^3.1.1",
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
     "@storybook/addon-a11y": "^6.5.10",
     "@storybook/addon-actions": "^6.1.16",
@@ -43,6 +44,7 @@
     "@storybook/addon-links": "^6.1.16",
     "@storybook/builder-vite": "^0.2.2",
     "@storybook/react": "^6.1.16",
+    "@storybook/testing-react": "^1.3.0",
     "@testing-library/dom": "^8.13.0",
     "@types/jest": "^28.1.8",
     "@types/node": "^17.0.35",
